+++ src/game-ld53/score.ts
