+++ src/game-ld53/ship.ts
