+++ src/game-ld54/ore.ts
