+++ src/game-ld54/space-path.ts
