+++ src/game-ld54/space-suit-controller.ts
